split service classes into endpoints and models
