+++ RechargeSharp.v2021-11.Tests/Configuration/DependencyInjection/RechargeSharpDependencyInjectionTests.cs
@@ -1,5 +1,4 @@
 using System;
-using System.ComponentModel.DataAnnotations;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
