+++ RechargeSharp.v2021-11/Endpoints/Subscriptions/SubscriptionService.Models.cs
@@ -1,87 +1,7 @@
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
-using RechargeSharp.v2021_11.Utilities;
-using RechargeSharp.v2021_11.Utilities.Queries;
-
 namespace RechargeSharp.v2021_11.Endpoints.Subscriptions;
 
-public class SubscriptionService
+public partial class SubscriptionService
 {
-    private readonly ILogger<SubscriptionService> _logger;
-    private readonly IRechargeApiCaller _rechargeApiCaller;
-
-    public SubscriptionService(ILogger<SubscriptionService> logger, IRechargeApiCaller rechargeApiCaller)
-    {
-        _logger = logger;
-        _rechargeApiCaller = rechargeApiCaller;
-    }
-
-    public async Task<CreateSubscriptionTypes.Response> CreateSubscriptionAsync(CreateSubscriptionTypes.Request request)
-    {
-        var requestUri = $"/subscriptions";
-        var responseJson = await _rechargeApiCaller.PostAsync<CreateSubscriptionTypes.Request, CreateSubscriptionTypes.Response> (request, requestUri);
-        return responseJson;
-    }
-    
-    
-    public async Task<GetSubscriptionTypes.Response> GetSubscriptionAsync(int subscriptionId)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}";
-        var responseJson = await _rechargeApiCaller.GetAsync<GetSubscriptionTypes.Response> (requestUri);
-        return responseJson;
-    }
-    
-    public async Task<UpdateSubscriptionTypes.Response> UpdateSubscriptionAsync(int subscriptionId, UpdateSubscriptionTypes.Request request)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}";
-        var responseJson = await _rechargeApiCaller.PutAsync<UpdateSubscriptionTypes.Request, UpdateSubscriptionTypes.Response>(request,requestUri);
-        return responseJson;
-    }
-    
-    public async Task<DeleteSubscriptionTypes.Response> DeleteSubscriptionAsync(int subscriptionId)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}";
-        await _rechargeApiCaller.DeleteAsync(requestUri);
-        return new DeleteSubscriptionTypes.Response();
-    }
-    
-    public async Task<ListSubscriptionsTypes.Response> ListSubscriptionsAsync(ListSubscriptionsTypes.Request request)
-    {
-        var queryString = ObjectToQueryStringSerializer.SerializeObjectToQueryString(request);
-        var requestUri = $"/subscriptions{queryString.Value}";
-        var responseJson = await _rechargeApiCaller.GetAsync<ListSubscriptionsTypes.Response>(requestUri);
-        return responseJson;
-    }
-    
-    public async Task<ChangeSubscriptionsNextChargeDateTypes.Response> ChangeSubscriptionsNextChargeDateAsync(int subscriptionId, ChangeSubscriptionsNextChargeDateTypes.Request request)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}/set_next_charge_date";
-        var responseJson = await _rechargeApiCaller.PostAsync<ChangeSubscriptionsNextChargeDateTypes.Request, ChangeSubscriptionsNextChargeDateTypes.Response>(request, requestUri);
-        return responseJson;
-    }
-    
-    public async Task<ChangeSubscriptionsAddressTypes.Response> ChangeSubscriptionsAddressAsync(int subscriptionId, ChangeSubscriptionsAddressTypes.Request request)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}/change_address";
-        var responseJson = await _rechargeApiCaller.PostAsync<ChangeSubscriptionsAddressTypes.Request, ChangeSubscriptionsAddressTypes.Response>(request, requestUri);
-        return responseJson;
-    }
-    
-    public async Task<CancelSubscriptionTypes.Response> CancelSubscriptionAsync(int subscriptionId, CancelSubscriptionTypes.Request request)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}/cancel";
-        var responseJson = await _rechargeApiCaller.PostAsync<CancelSubscriptionTypes.Request, CancelSubscriptionTypes.Response>(request, requestUri);
-        return responseJson;
-    }
-    
-    public async Task<ActivateSubscriptionTypes.Response> ActivateSubscriptionAsync(int subscriptionId)
-    {
-        var requestUri = $"/subscriptions/{subscriptionId}/activate";
-        var responseJson = await _rechargeApiCaller.PostAsync<ActivateSubscriptionTypes.Response>(requestUri);
-        return responseJson;
-    }
-    
-
     public static class SharedSubscriptionTypes
     {
         public record Property(
@@ -92,15 +12,15 @@ public class SubscriptionService
         public record ExternalVariantId(
             string Ecommerce
         );
-        
+
         public record ExternalProductId(
             string Ecommerce
         );
-        
+
         public record AnalyticsData(
             IReadOnlyList<object> UtmParams
         );
-        
+
         public record Subscription(
             int Id,
             int AddressId,
@@ -157,7 +77,7 @@ public class SubscriptionService
             SharedSubscriptionTypes.ExternalVariantId ExternalVariantId,
             string? Status
         );
-        
+
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
     }
 
@@ -165,7 +85,7 @@ public class SubscriptionService
     {
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
     }
-    
+
     public static class UpdateSubscriptionTypes
     {
         public record Request(
@@ -187,7 +107,7 @@ public class SubscriptionService
             string? UseExternalVariantDefaults,
             string? VariantTitle
         );
-        
+
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
     }
 
@@ -211,33 +131,36 @@ public class SubscriptionService
             string? Status,
             DateTime? UpdatedAtMax,
             DateTime? UpdatedAtMin
-            );
-        
+        );
+
         public record Response(
             string? NextCursor,
             string? PreviousCursor,
             IReadOnlyList<SharedSubscriptionTypes.Subscription> Subscriptions
-            );
+        );
     }
 
     public static class ChangeSubscriptionsNextChargeDateTypes
     {
         public record Request(DateOnly Date);
+
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
     }
 
     public static class ChangeSubscriptionsAddressTypes
     {
         public record Request(int AddressId, DateOnly? NextChargeScheduledAt);
+
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
     }
-    
+
     public static class CancelSubscriptionTypes
     {
         public record Request(string CancellationReason, string? CancellationReasonComments, bool SendEmail);
+
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
     }
-    
+
     public static class ActivateSubscriptionTypes
     {
         public record Response(SharedSubscriptionTypes.Subscription Subscription);
